requirements.txt
Flask==3.0.3
Pillow==10.0.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
Werkzeug==3.0.1